[FIX] push and clone

+++ git-lib/src/objects_database.rs
@@ -153,7 +153,12 @@ impl ObjectsDatabase {
         if let Err(error) = fs::create_dir_all(parent_path) {
             return Err(CommandError::FileOpenError(error.to_string()));
         };
+        // si existe el path, no se guarda
+        if std::path::Path::new(&path).exists() {
+            return Ok(hash_str);
+        }
         let Ok(mut file) = File::create(&path) else {
+            println!("Error al crear archivo para escrituraafafafa");
             return Err(CommandError::FileOpenError(
                 "Error al abrir archivo para escritura".to_string(),
             ));
